deployment/requirements.txt
fastapi==0.117.1
matplotlib==3.10.6
opencv_python_headless==4.10.0.84
pandas==2.3.2
Pillow==11.3.0
resnet==0.1
scikit_learn==1.7.2
scipy==1.16.2
scikit_image==0.24.0
timm==1.0.20
tqdm==4.66.5
uvicorn==0.36.0
tensorboard==2.18.0
python-multipart
